Create Repository module
