+++ test/test_helper.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "retest"
-
+require "byebug"
 require "minitest/autorun"
